Added Score matrix calculation, Needleman-Wunsch take O(mn) space
The current implementation calculates the score by Needleman-Wunsch algorithm,
this algorithm takes O(mn) space which in turn will throw a HeapSpace Exception
around 1000 seq long alignments on default java heap. So, it works but is very
limited on the size of the sequences that can be aligned. After some research
there's a Hischberg's algorithm which takes less space by divide and conquer
strategy over Needleman-Wunsch.